+++ Sensors/SMCSuperIO/ITEDevice.hpp
@@ -95,8 +95,8 @@ namespace ITE {
 		/**
 		 *  Device access
 		 */
-		uint8_t readByte(uint8_t reg);
-		void writeByte(uint8_t reg, uint8_t value);
+		uint8_t readByte(uint16_t reg);
+		void writeByte(uint16_t reg, uint8_t value);
 
 		/**
 		 *  Overrides
